Update patch set 2
Patch Set 2: Code-Review-1

I'm not sure if you're synced with ToT, but we're in the process of phasing out the code in CameraSettings.  You'll noticed that getSupportedHighestVideoQuality is no longer used.  Instead we set the video quality to default and allow the user to modify it via the settings menu.

Patch-set: 2
Reviewer: Gerrit User 1036513 <1036513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1