Update patch set 1
Patch Set 1:

This should be fixed already by https://android-review.googlesource.com/#/c/82776/
When LOCAL_SDK_VERSION is specified the module is built against the NDK which provides prebuilts of liblog, etc.  LOCAL_SDK_VERSION was only set for arm, mips, and x86, which caused arm64 to try to link directly against the locally built version of liblog but without the dependency.

Patch-set: 1