Fix the crash issue caused by resolution switch
On low performance device, with "Auto-rotate" off, and the display screen is portrait.
open the Camera2 to launch preview activity,
click camera seettings button and then return to the preview activity,
Sometimes crash happened, the preview is in landscape and
the seettings interface is in portrait when crashed.
According to the crash log "Attempt to invoke virtual method
'void com.android.camera.settings.SettingsManager.addListener
(com.android.camera.settings.SettingsManager$OnSettingChangedListener)' on a null object reference",
we can find that Null pointer variable caused the crash.

This issue is analyzed as below: when returned back to the preview activity,
CameraActivity has gone through 2 periods:
onPause -> onDestroyTasks -> OnCreateTasks -> resume, Inside the resume() function,
"openCameraAndStartPreview()" in CaptureModule.java will be called,
and a new thread may be started in callback function "onCameraOpened".

As mentioned aove, in this case there will be 2 lifecycle calls,
At some extreme timing, the following sequence of operations will occur:
onPause1 -> onDestroyTasks1 -> OnCreateTasks1 -> resume1(start thread1) ->
onPause2 -> onDestroyTasks2 -> OnCreateTasks2 -> thread1 running, and crash.
thread1 started at the first cycle, and the related variables have been destroyed later,
but the thread1 is actually running in the second cycle.
Therefore, when the thread1 function is executed further down,
"mAppController.getSettingsManager()" is null, which caused the crash.

To fix this issue, add a flag to ensure that the
thread1 can not be executed after the first cycle is destroyed

Change-Id: I5a970cbe85de6442c2b4f0960b6a189304c72477
Signed-off-by: Anle Pan <anle.pan@nxp.com>