MA-21337 Fix Camera Intents test fail issue
The CtsVerifier case failed because of the location check,
the case requires no location when taking picture.
During the test 3 of 5, case will trigger camera2
activity to test, even not allow the location permission,
the GPS tag was default set in initTagInfo when CompreseeImageToJpeg.
delete the GPS tag when GPS is not available.

Change-Id: I6e7e8b7fbe5cb7412679ed0293ef4d6503840069
Signed-off-by: Anle Pan <anle.pan@nxp.com>